Controladores y validaciones de tags, category, user rates, partner rates, tags branches

+++ app/Http/Controllers/BranchController.php
@@ -10,11 +10,17 @@ use App\Partner;
 use App\Company;
 use App\Tag;
 use App\TagBranch;
-
 use Validator;
+use JWTAuth;
 
 class BranchController extends Controller
 {
+	
+	public function __construct(){
+        $this->middleware('jwt.auth:partner', ['only' => ['store','show','update','destroy','tags','tagStore','tagUpdate','tagDestroy']]);
+        $this->middleware('default.headers');
+    }
+	
     /**
      * Display a listing of the resource.
      *
@@ -22,12 +28,8 @@ class BranchController extends Controller
      */
     public function index()
     {
-		$branches = Branch::with('company.partner')->get();       
-		return response()->json($branches->all(),200);
-		
-		//PARA GENERAR LA VISTA Y HACER PRUEBAS
-		//$branches = Branch::all();
-		//return view('branch.index')->with('branches',$branches);
+		/*$branches = Branch::with('company.partner')->get();       
+		return response()->json($branches->all(),200);*/
 		
     }
 
@@ -38,9 +40,7 @@ class BranchController extends Controller
      */
     public function create()
     {
-		$companies = Company::all();
-		return view('branch.create')->with('companies',$companies);
-		
+		//	
 		
     }
 
@@ -69,25 +69,35 @@ class BranchController extends Controller
 		
 		if(!is_null($company)){
 			
-			//SE UNA INSTANCIA DE BRANCH
-			$branch = new Branch;
-			$branch->company_id = $company_id ;
-			$branch->address = $request->address;
-			$branch->phone = $request->phone;
-			$branch->latitude = $request->latitude;
-			$branch->longitude = $request->longitude;
-			$branch->state_id = 1;
-			$branch->schedule = $request->schedule;
+			$partnerRequested = \Auth::User();
 			
-			$row = $branch->save();
-			
-			if($row != false){
-				$response = ['data' => $branch,'code' => 200,'message' => 'Branch was created succefully'];
-				return response()->json($response,200);
+			//SE VERIFICA QUE EL PARTNER QUE HIZO LA PETICION SOLO PUEDA GUARDAR BRANCHES EN SUS COMPANIES
+			if($partnerRequested->id == $company->partner_id){
+				
+				//SE UNA INSTANCIA DE BRANCH
+				$branch = new Branch;
+				$branch->company_id = $company_id ;
+				$branch->address = $request->address;
+				$branch->phone = $request->phone;
+				$branch->latitude = $request->latitude;
+				$branch->longitude = $request->longitude;
+				$branch->state_id = 1;
+				$branch->schedule = $request->schedule;
+				
+				$row = $branch->save();
+				
+				if($row != false){
+					$response = ['data' => $branch,'code' => 200,'message' => 'Branch was created succefully'];
+					return response()->json($response,200);
+				}else{
+					$response = ['error' => 'It has occurred an error trying to save the branch','code' => 404];
+					return response()->json($response,404);
+				}
 			}else{
-				$response = ['error' => 'It has occurred an error trying to save the branch','code' => 404];
-				return response()->json($response,404);
+				$response = ['error'   => 'Unauthorized','code' => 403];
+				return response()->json($response, 403);
 			}
+					
 		}else{
 			//EN DADO CASO QUE EL ID DE LA COMPANY NO SE HALLA ENCONTRADO
 			$response = ['error' => 'Company does not exist','code' => 422];
@@ -104,10 +114,29 @@ class BranchController extends Controller
      */
     public function show($id)
     {
-        $branch = Branch::find($id);
+		$partnerRequested = \Auth::User();
+		
+		//SE OBTIENE LA BRANCH SOLICITIDA JUNTO CON LA COMPANY QUE LE PERTENECE
+        $branch = Branch::with('company')->where('id','=',$id)->first();
+		
+		//SE VALIDA QUE EXISTA LA BRANCH
 		if(!is_null($branch)){
-			$response = ['code' => 200,'data' => $branch];
-			return response()->json($response,200);
+			
+			//SE OBTIENE LA COMPANY DE LA BRANCH
+			$company = $branch->company;		
+			
+			//SE VERIFICA QUE EL PARTNER QUE HIZO LA PETICION SOLO PUEDA OBTENER INFO DE SUS BRANCHES
+			if($partnerRequested->id == $company->partner_id){
+				
+				$response = ['code' => 200,'data' => $branch];
+				return response()->json($response,200);
+				
+			}else{
+				
+				$response = ['error'   => 'Unauthorized','code' => 403];
+				return response()->json($response, 403);
+			}		
+			
 		}else{
 			$response = ['error' => 'Branch does no exist','code' => 422];
 			return response()->json($response,422);
@@ -122,14 +151,7 @@ class BranchController extends Controller
      */
     public function edit($id)
     {
-		$branch = Branch::find($id);	
-		if(!is_null($branch)){
-			return view('branch.edit')->with('branch',$branch);
-		}else{
-			$response = ['error' => 'Branch does no exist','code' => 404];
-			return response()->json($response,404,[],JSON_PRETTY_PRINT);
-		}
-        
+        //
     }
 
     /**
@@ -141,6 +163,8 @@ class BranchController extends Controller
      */
     public function update(Request $request, $id)
     {
+		$partnerRequested = \Auth::User();
+		
         $messages = Branch::getMessages();
 		$validation = Branch::getValidations();
 		
@@ -152,25 +176,38 @@ class BranchController extends Controller
 			return response()->json($response,422);
 		}
 		
-		$branch = Branch::find($id);
+		//SE OBTIENE LA BRANCH SOLICITIDA JUNTO CON LA COMPANY QUE LE PERTENECE
+        $branch = Branch::with('company')->where('id','=',$id)->first();
 		
 		//SE VALIDA QUE LA BRANCH EXISTA
 		if(!is_null($branch)){
 			
-			//SE GUARDAN EN UN ARREGLO LOS CAMPOS QUE SE PUEDEN ACTUALIZAR Y SE IGUALAN A LOS QUE VIENEN POR LA PETICION
-			$fields = ['address' => $request->address,'phone' => $request->phone,'schedule' => $request->schedule,
-			'latitude' => $request->latitude, 'longitude' => $request->longitude];
+			//SE OBTIENE LA COMPANY DE LA BRANCH
+			$company = $branch->company;
 			
-			$row = Branch::where('id','=',$id)->update($fields);
-			
-			//SE VALIDA QUE SE HALLA ACTUALIZADO EL REGISTRO
-			if($row != false){
-				$response = ['data' => $branch,'code' => 200,'message' => 'Branch was updated succefully'];
-				return response()->json($response,200);
+			//SE VERIFICA QUE EL PARTNER QUE HIZO LA PETICION SOLO PUEDA ACTUALIZAR SUS BRANCHES
+			if($partnerRequested->id == $company->partner_id){
+				
+				//SE GUARDAN EN UN ARREGLO LOS CAMPOS QUE SE PUEDEN ACTUALIZAR Y SE IGUALAN A LOS QUE VIENEN POR LA PETICION
+				$fields = ['address' => $request->address,'phone' => $request->phone,'schedule' => $request->schedule,
+				'latitude' => $request->latitude, 'longitude' => $request->longitude];
+				
+				$row = Branch::where('id','=',$id)->update($fields);
+				
+				//SE VALIDA QUE SE HALLA ACTUALIZADO EL REGISTRO
+				if($row != false){
+					$response = ['data' => $branch,'code' => 200,'message' => 'Branch was updated succefully'];
+					return response()->json($response,200);
+				}else{
+					$response = ['error' => 'It has occurred an error trying to update the branch','code' => 404];
+					return response()->json($response,404);
+				}
+				
 			}else{
-				$response = ['error' => 'It has occurred an error trying to update the branch','code' => 404];
-				return response()->json($response,404);
+				$response = ['error'   => 'Unauthorized','code' => 403];
+				return response()->json($response, 403);
 			}
+						
 		}else{
 			//EN DADO CASO QUE EL ID DE BRANCH NO SE HALLA ENCONTRADO
 			$response = ['error' => 'Branch does not exist','code' => 422];
@@ -187,20 +224,33 @@ class BranchController extends Controller
      */
     public function destroy($id)
     {
-        $branch = Branch::find($id);
+		$partnerRequested = \Auth::User();
+		
+        //SE OBTIENE LA BRANCH SOLICITIDA JUNTO CON LA COMPANY QUE LE PERTENECE
+        $branch = Branch::with('company')->where('id','=',$id)->first();
 
 		if(!is_null($branch)){
 			
-			//SE BORRAR LA BRANCH
-			$rows = $branch->delete();
+			//SE OBTIENE LA COMPANY DE LA BRANCH
+			$company = $branch->company;
 			
-			if($rows > 0){
-				$response = ['code' => 200,'message' => "Branch was deleted succefully"];
-				return response()->json($response,200);
+			//SE VERIFICA QUE EL PARTNER QUE HIZO LA PETICION SOLO PUEDA ELIMINAR SUS BRANCHES
+			if($partnerRequested->id == $company->partner_id){
+				
+				//SE BORRAR LA BRANCH
+				$rows = $branch->delete();
+				
+				if($rows > 0){
+					$response = ['code' => 200,'message' => "Branch was deleted succefully"];
+					return response()->json($response,200);
+				}else{
+					$response = ['error' => 'It has occurred an error trying to delete the branch','code' => 404];
+					return response()->json($response,404);
+				}
 			}else{
-				$response = ['error' => 'It has occurred an error trying to delete the branch','code' => 404];
-				return response()->json($response,404);
-			}			
+				$response = ['error'   => 'Unauthorized','code' => 403];
+				return response()->json($response, 403);
+			}								
 			
 		}else{
 			//EN DADO CASO QUE EL ID DE LA BRANCH NO SE HALLA ENCONTRADO
@@ -217,18 +267,33 @@ class BranchController extends Controller
 	*/
 	public function tags($id){
 		
-		$branch = Branch::find($id);
+		$partnerRequested = \Auth::User();
+		
+		//SE OBTIENE LA BRANCH SOLICITIDA JUNTO CON LA COMPANY QUE LE PERTENECE
+        $branch = Branch::with('company')->where('id','=',$id)->first();
+
 		if(!is_null($branch)){
 			
-			$tags = \DB::table('tags_branches')
-			->join('tags','tags_branches.tag_id','=','tags.id')
-			->where('branch_id','=',$id)
-			->select('tags.*')
-			->get();
+			//SE OBTIENE LA COMPANY DE LA BRANCH
+			$company = $branch->company;
 			
-			$response = ['data' => $tags,'code' => 200];
-			
-			return response()->json($response,200);
+			//SE VERIFICA QUE EL PARTNER QUE HIZO LA PETICION SOLO PUEDA OBTENER INFO DE SUS BRANCHES
+			if($partnerRequested->id == $company->partner_id){
+				
+				$tags = \DB::table('tags_branches')
+				->join('tags','tags_branches.tag_id','=','tags.id')
+				->where('branch_id','=',$id)
+				->select('tags.id','tags.name','tags.description')
+				->get();
+				
+				$response = ['data' => $tags,'code' => 200];
+				
+				return response()->json($response,200);
+				
+			}else{
+				$response = ['error'   => 'Unauthorized','code' => 403];
+				return response()->json($response, 403);
+			}						
 			
 		}else{
 			//EN DADO CASO QUE EL ID DE BRANCH NO SE HALLA ENCONTRADO
@@ -243,6 +308,8 @@ class BranchController extends Controller
 	* la jerarquía, ya que un tag puede pertenecer a una branch.
 	*/
 	public function tagStore(Request $request){
+		
+		$partnerRequested = \Auth::User();
 		
 		$messages = TagBranch::getMessages();
 		$validation = TagBranch::getValidations();
@@ -259,28 +326,40 @@ class BranchController extends Controller
 		$branch_id = $request->branch_id;
 		$tag_id = $request->tag_id;
 		
-		$branch = Branch::find($branch_id);
+		//SE OBTIENE LA BRANCH SOLICITIDA JUNTO CON LA COMPANY QUE LE PERTENECE
+        $branch = Branch::with('company')->where('id','=',$branch_id)->first();
+		
 		$tag = Tag::find($tag_id);
 		
 		//SE VERIFICA QUE LA BRANCH Y CATEGORY EXISTAN
 		if(!is_null($branch) && !is_null($tag)){
 			
-			//SE GUARDA EL TAG QUE LE PERTENECE A LA BRANCH
-			$row = \DB::table('tags_branches')->insert(
-				[
-					'tag_id' => $tag->id,
-					'branch_id' => $branch->id
-				]
-			);
+			//SE OBTIENE LA COMPANY DE LA BRANCH
+			$company = $branch->company;
 			
-			//SI LAS ROWS AFECTADAS SON IGUAL A 1 O MAS ENTONCES SI SE GUARDO
-			if($row != false){
-				$response = ['code' => 200,'message' => 'Tag was created succefully'];
-				return response()->json($response,200);
+			//SE VERIFICA QUE EL PARTNER QUE HIZO LA PETICION SOLO PUEDA GUARDAR TAGS DE SUS BRANCHES
+			if($partnerRequested->id == $company->partner_id){
+				
+				//SE GUARDA EL TAG QUE LE PERTENECE A LA BRANCH
+				$row = \DB::table('tags_branches')->insert(
+					[
+						'tag_id' => $tag->id,
+						'branch_id' => $branch->id
+					]
+				);
+				
+				//SI LAS ROWS AFECTADAS SON IGUAL A 1 O MAS ENTONCES SI SE GUARDO
+				if($row != false){
+					$response = ['code' => 200,'message' => 'Tag was created succefully'];
+					return response()->json($response,200);
+				}else{
+					$response = ['error' => 'It has occurred an error trying to update the tag','code' => 404];
+					return response()->json($response,404);
+				}
 			}else{
-				$response = ['error' => 'It has occurred an error trying to update the tag','code' => 404];
-				return response()->json($response,404);
-			}
+				$response = ['error'   => 'Unauthorized','code' => 403];
+				return response()->json($response, 403);
+			}					
 			
 		}else{
 			//EN DADO CASO QUE EL ID DE CATEGORY NO SE HALLA ENCONTRADO
@@ -297,6 +376,8 @@ class BranchController extends Controller
 	*/
 	public function tagUpdate(Request $request, $id)
     {
+		$partnerRequested = \Auth::User();
+		
         $messages = TagBranch::getMessages();
 		$validation = TagBranch::getValidations();
 		
@@ -307,29 +388,44 @@ class BranchController extends Controller
 		//SE VERIFICA SI ALGUN CAMPO NO ESTA CORRECTO
 		if($v->fails()){	
 			return response()->json($response,460,[],JSON_PRETTY_PRINT);
-		}
+		}		
 		
 		$tag_id = $request->tag_id;
-		//$tag = DB::table('tags_branches')->where('id','=',$id)->get();
-		$tag = \DB::table('tags_branches')->where('id','=',$id)->first();
+		
+		//SE RELACIONA EL TAG CON LA BRANCH QUE PERTENECE Y LA BRANCH SE RELACIONA CON LA COMPANY QUE PERTENECE 
+		$tag = \DB::table('tags_branches')
+			->join('branches','tags_branches.branch_id','=','branches.id')
+			->join('companies','branches.company_id','=','companies.id')
+			->where('tags_branches.id','=',$id)
+			->select('companies.id AS company_id')
+			->first();
 		
 		//SE VALIDA QUE EL TAG A ACTUALIZAR EXISTA
-		//if(count($tag) > 0){
 		if(!is_null($tag)){
 			
-			//SE GUARDAN EN UN ARREGLO LOS CAMPOS QUE SE PUEDEN ACTUALIZAR Y SE IGUALAN A LOS QUE VIENEN POR LA PETICION
-			$fields = ['tag_id' => $tag_id];
+			//SE OBTIENE LA COMPANY DEL TAG	
+			$company = Company::find($tag->company_id);
 			
-			$row = \DB::table('tags_branches')->where('id','=',$id)->update($fields);
-			
-			//SI LAS ROWS AFECTADAS SON IGUAL A 1 O MAS ENTONCES SI SE GUARDO
-			if($row != false){
-				$response = ['code' => 200,'message' => 'Tag was updated succefully'];
-				return response()->json($response,200);
+			//SE VERIFICA QUE EL PARTNER QUE HIZO LA PETICION SOLO PUEDA ACTUALIZAR SUS TAGS
+			if($partnerRequested->id == $company->partner_id){
+				
+				//SE GUARDAN EN UN ARREGLO LOS CAMPOS QUE SE PUEDEN ACTUALIZAR Y SE IGUALAN A LOS QUE VIENEN POR LA PETICION
+				$fields = ['tag_id' => $tag_id];
+				
+				$row = \DB::table('tags_branches')->where('id','=',$id)->update($fields);
+				
+				//SI LAS ROWS AFECTADAS SON IGUAL A 1 O MAS ENTONCES SI SE GUARDO
+				if($row != false){
+					$response = ['code' => 200,'message' => 'Tag was updated succefully'];
+					return response()->json($response,200);
+				}else{
+					$response = ['error' => 'It has occurred an error trying to update the tag','code' => 404];
+					return response()->json($response,404);
+				}
 			}else{
-				$response = ['error' => 'It has occurred an error trying to update the tag','code' => 404];
-				return response()->json($response,404);
-			}
+				$response = ['error'   => 'Unauthorized','code' => 403];
+				return response()->json($response, 403);
+			}						
 			
 		}else{
 			//EN DADO CASO QUE EL ID DEL TAG NO SE HALLA ENCONTRADO
@@ -346,26 +442,39 @@ class BranchController extends Controller
 	*/
 	public function tagDestroy($id){
 		
-		//SE OBTIENE EL TAG EN FORMA DE OBJETO
-		$tag = \DB::table('tags_branches')->where('id','=',$id)->first();		
+		$partnerRequested = \Auth::User();
+		
+		//SE RELACIONA EL TAG CON LA BRANCH QUE PERTENECE Y LA BRANCH SE RELACIONA CON LA COMPANY QUE PERTENECE 
+		$tag = \DB::table('tags_branches')
+			->join('branches','tags_branches.branch_id','=','branches.id')
+			->join('companies','branches.company_id','=','companies.id')
+			->where('tags_branches.id','=',$id)
+			->select('companies.id AS company_id')
+			->first();	
 
 		//SE VALIDA QUE EL TAG EXISTA
 		if(!is_null($tag)){
 			
-			//SE BORRA EL TAG
-	
-			//$row = \DB::table('tags_branches')->where('id','=',$tag->id)->delete();
+			//SE OBTIENE LA COMPANY DEL TAG	
+			$company = Company::find($tag->company_id);
 			
-			$fields = ['deleted_at' => date('Y-M-d hh:mm:ss',time())];
-			$row = \DB::table('tags_branches')->where('id','=',$tag->id)->update($fields);
-			
-			if($row != false){
-				$response = ['code' => 200,'message' => "Tag was deleted succefully"];
-				return response()->json($response,200);
+			//SE VERIFICA QUE EL PARTNER QUE HIZO LA PETICION SOLO PUEDA ELIMINAR SUS TAGS DE SUS BRANCHES
+			if($partnerRequested->id == $company->partner_id){
+				
+				$fields = ['deleted_at' => date('Y-M-d hh:mm:ss',time())];
+				$row = \DB::table('tags_branches')->where('id','=',$id)->update($fields);
+				
+				if($row != false){
+					$response = ['code' => 200,'message' => "Tag was deleted succefully"];
+					return response()->json($response,200);
+				}else{
+					$response = ['error' => 'It has occurred an error trying to delete the tag','code' => 404];
+					return response()->json($response,404);
+				}
 			}else{
-				$response = ['error' => 'It has occurred an error trying to delete the tag','code' => 404];
-				return response()->json($response,404);
-			}			
+				$response = ['error'   => 'Unauthorized','code' => 403];
+				return response()->json($response, 403);
+			}									
 			
 		}else{
 			//EN DADO CASO QUE EL ID DE LA BRANCH NO SE HALLA ENCONTRADO
